requirements.txt
pytest==6.2.5
Appium-Python-Client==2.2.0